Refactor listicle code
